feat: implement authentication routes including login, logout, refresh, and signup; refactor connection deletion logic
