--- src/server/routes/api/auth/logout/post.ts
@@ -1,14 +1,0 @@
-import {FastifyPluginAsync} from 'fastify';
-import SQL from 'sql-template-strings';
-
-const plugin: FastifyPluginAsync = async server => {
-  server.post('', {onRequest: server.authenticate}, async (request, reply) => {
-    await server.db.run(
-      SQL`DELETE FROM connections WHERE id = ${request.connection}`,
-    );
-
-    return reply.clearCookie('refreshToken').code(204).send();
-  });
-};
-
-export default plugin;
