+++ src/server/routes/api/healthcheck.ts
@@ -1,7 +1,7 @@
 import {FastifyPluginAsync} from 'fastify';
 
 const plugin: FastifyPluginAsync = async server => {
-  server.get('', () => ({
+  server.get('/healthcheck', () => ({
     status: 'ok',
     uptime: process.uptime(),
   }));
